requirements.txt
Flask==3.1.0
protobuf==5.29.0
requests==2.31.0
gunicorn==20.1.0
google-generativeai
python-pptx

